Base: Setup Project-User relationship, notifications

+++ app/Http/Controllers/ProjectController.php
@@ -2,14 +2,22 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\Projects\CreateProjectRequest;
-use App\Http\Requests\Projects\CreateProjectsRequest;
-use App\Http\Requests\Projects\UpdateProjectRequest;
+use App\Models\User;
 use App\Models\Project;
 use Illuminate\Http\Request;
 use Spatie\QueryBuilder\QueryBuilder;
 use App\Http\Resources\ProjectsResource;
 use App\Http\Resources\ProjectsCollection;
+use Illuminate\Support\Facades\Notification;
+use App\Http\Requests\Projects\InviteUserRequest;
+use App\Http\Requests\Projects\CreateProjectRequest;
+use App\Http\Requests\Projects\UpdateProjectRequest;
+use App\Http\Requests\Projects\ChangeProjectOwnershipRequest;
+use App\Notifications\NotifyInvitedUsers;
+use App\Notifications\NotifyRevokedUsers;
+use App\Notifications\ProjectDeleteNotification;
+use App\Notifications\ProjectOwnerShipChange;
+use Illuminate\Database\Eloquent\Collection;
 
 class ProjectController extends Controller
 {
@@ -27,16 +35,7 @@ class ProjectController extends Controller
         ])->jsonPaginate();
         return new ProjectsCollection($projects);
     }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-    }
-
+    
     /**
      * Store a newly created resource in storage.
      *
@@ -47,6 +46,7 @@ class ProjectController extends Controller
     {
         $project = Project::create([
             'name' => $request->input('data.attributes.name'),
+            'user_id' => $request->input('data.attributes.user_id')
         ]);
         return (new ProjectsResource($project))->response()->header('Location', route('projects.show', ['project' => $project]));
     }
@@ -59,19 +59,44 @@ class ProjectController extends Controller
      */
     public function show(Project $project)
     {
-        return new ProjectsResource($project);
+        return [Collect($project->invitees)];
+        // return new ProjectsResource($project);
     }
-
+    
     /**
-     * Show the form for editing the specified resource.
+     * Project creator can invite other users.
      *
      * @param  \App\Models\Project  $project
      * @return \Illuminate\Http\Response
      */
-    public function edit(Project $project)
+    public function invite(Project $project, InviteUserRequest $request)
     {
-        //
+        if ($request->user()->cannot('invite', $project)) {
+            abort(403, 'You are not the owner of this project');
+        }
+
+        $project->invitees()->attach($request->input('data.attributes.id'));
+
+        if ($project->invitees) {
+            Notification::send($project->invitees, new NotifyInvitedUsers($project));
+        }
+        return response(null, 201);
     }
+    
+    public function revoke(Project $project, InviteUserRequest $request)
+    {
+        if ($request->user()->cannot('revoke', $project)) {
+            abort(403, 'You are not the owner of this project');
+        }
+        
+        $project->invitees()->detach($request->input('data.attributes.id'));
+
+        if ($project->invitees) {
+            Notification::send($project->invitees, new NotifyRevokedUsers($project));
+        }
+        return response(null, 204);
+    }
+
 
     /**
      * Update the specified resource in storage.
@@ -85,6 +110,28 @@ class ProjectController extends Controller
         $project->update($request->input('data.attributes'));
         return new ProjectsResource($project);
     }
+    
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Models\Project  $project
+     * @return \Illuminate\Http\Response
+     */
+    public function change_ownership(ChangeProjectOwnershipRequest $request, Project $project)
+    {
+        if ($request->user()->cannot('change_ownership', $project)) {
+            abort(403, 'You are not the owner of this project');
+        }
+
+        $project->update($request->input('data.attributes'));
+        $user = User::find($request->input('data.attributes.user_id'));
+
+        if ($user) {
+            $user->notify(new ProjectOwnerShipChange($project));
+        }
+        return response(null, 204);
+    }
 
     /**
      * Remove the specified resource from storage.
@@ -92,8 +139,14 @@ class ProjectController extends Controller
      * @param  \App\Models\Project  $project
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Project $project)
+    public function destroy(Request $request, Project $project)
     {
+        if ($request->user()->cannot('delete', $project)) {
+            abort(403, 'You are not the owner of this project');
+        }
+        
+        // Notification::send($project->invitees, new ProjectDeleteNotification($project));
+        
         $project->delete();
         return response(null, 204);
     }
